requirements.txt
cycler==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.2.1
numpy==1.20.1
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.1
python-dateutil==2.8.1
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
pillow==8.1.2
opencv-python==4.5.1.48